+++ src/main.rs
@@ -49,7 +49,6 @@ macro_rules! command_list {
 
 ## Search & filter
     search      Search for (or replace) patterns in CSV data
-    bisect      Search for value using binary search in sorted CSV data
     grep        Coarse but fast filtering of CSV data
     filter      Only keep some CSV rows based on an evaluated expression
     head        First rows of CSV file
@@ -57,6 +56,7 @@ macro_rules! command_list {
     slice       Slice rows of CSV file
     top         Find top rows of a CSV file according to some column
     sample      Randomly sample CSV data
+    bisect      Binary search on sorted CSV data
 
 ## Sort & deduplicate
     sort        Sort CSV data